pyproject.toml
[project]
name = "gaia"
version = "0.1.0"
description = "GAIA Benchmark Agent with Multi-Modal File Processing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core LangChain
    "langchain>=1.1.3",
    "langchain-huggingface>=1.1.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.4",
    
    # HuggingFace
    "huggingface-hub>=0.36.0",
    "transformers>=4.46.0",
    "accelerate>=1.0.0",
    
    # Vector store
    "sentence-transformers>=5.2.0",
    "supabase>=2.25.1",
    
    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    
    # Data processing (polars)
    "polars>=1.0.0",
    "xlsx2csv>=0.8.0",
    
    # Science
    "biopython>=1.82",
    "numpy>=2.0.0",
    
    # Image
    "pillow>=10.0.0",
    
    # Audio
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    
    # Web tools
    "duckduckgo-search>=4.0.0",
    "tavily-python>=0.5.0",
    "wikipedia>=1.4.0",
    "arxiv>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "tqdm>=4.67.1",
    "gradio>=4.0.0",
    "ipykernel>=7.1.0",
]
